Add ooo-tmux script
This adds a ooo-tmux script that is used to establish a tmux session
with a TripleO development environment.  It expects that you are
running a TripleO deployment in VMs on a single host.  The script
can be executed directly on the hypervisor host, or remotely by
using the JUMPHOST option.